third_party/blink/web_tests/platform/generic/fast/dom/SelectorAPI/not-supported-namespace-in-selector-expected.txt

This tests that we throw a NAMESPACE_ERR when parsing a selector string for querySelector and querySelectorAll that contains a namespace.

PASS: document.querySelector('bbb|pre') throws: SyntaxError: Failed to execute 'querySelector' on 'Document': 'bbb|pre' is not a valid selector.
PASS: document.querySelectorAll('bbb|pre') throws: SyntaxError: Failed to execute 'querySelectorAll' on 'Document': 'bbb|pre' is not a valid selector.
PASS: document.body.webkitMatchesSelector('bbb|pre') throws: SyntaxError: Failed to execute 'webkitMatchesSelector' on 'Element': 'bbb|pre' is not a valid selector.
PASS: document.querySelector('*|pre') did not throw
PASS: document.querySelectorAll('*|pre') did not throw
PASS: document.body.webkitMatchesSelector('*|pre') did not throw
PASS: document.querySelector('|pre') did not throw
PASS: document.querySelectorAll('|pre') did not throw
PASS: document.body.webkitMatchesSelector('|pre') did not throw
PASS: document.querySelector('div bbb|pre') throws: SyntaxError: Failed to execute 'querySelector' on 'Document': 'div bbb|pre' is not a valid selector.
PASS: document.querySelectorAll('div bbb|pre') throws: SyntaxError: Failed to execute 'querySelectorAll' on 'Document': 'div bbb|pre' is not a valid selector.
PASS: document.body.webkitMatchesSelector('div bbb|pre') throws: SyntaxError: Failed to execute 'webkitMatchesSelector' on 'Element': 'div bbb|pre' is not a valid selector.
PASS: document.querySelector('div *|pre') did not throw
PASS: document.querySelectorAll('div *|pre') did not throw
PASS: document.body.webkitMatchesSelector('div *|pre') did not throw
PASS: document.querySelector('div |pre') did not throw
PASS: document.querySelectorAll('div |pre') did not throw
PASS: document.body.webkitMatchesSelector('div |pre') did not throw
PASS: document.querySelector('[bbb|name=value]') throws: SyntaxError: Failed to execute 'querySelector' on 'Document': '[bbb|name=value]' is not a valid selector.
PASS: document.querySelectorAll('[bbb|name=value]') throws: SyntaxError: Failed to execute 'querySelectorAll' on 'Document': '[bbb|name=value]' is not a valid selector.
PASS: document.body.webkitMatchesSelector('[bbb|name=value]') throws: SyntaxError: Failed to execute 'webkitMatchesSelector' on 'Element': '[bbb|name=value]' is not a valid selector.
PASS: document.querySelector('[*|name=value]') did not throw
PASS: document.querySelectorAll('[*|name=value]') did not throw
PASS: document.body.webkitMatchesSelector('[*|name=value]') did not throw
PASS: document.querySelector('[|name=value]') did not throw
PASS: document.querySelectorAll('[|name=value]') did not throw
PASS: document.body.webkitMatchesSelector('[|name=value]') did not throw
PASS: document.querySelector(':-webkit-any(bbb|pre)') throws: SyntaxError: Failed to execute 'querySelector' on 'Document': ':-webkit-any(bbb|pre)' is not a valid selector.
PASS: document.querySelector('div [bbb|name=value]') throws: SyntaxError: Failed to execute 'querySelector' on 'Document': 'div [bbb|name=value]' is not a valid selector.
PASS: document.querySelectorAll('div [bbb|name=value]') throws: SyntaxError: Failed to execute 'querySelectorAll' on 'Document': 'div [bbb|name=value]' is not a valid selector.
PASS: document.body.webkitMatchesSelector('div [bbb|name=value]') throws: SyntaxError: Failed to execute 'webkitMatchesSelector' on 'Element': 'div [bbb|name=value]' is not a valid selector.
PASS: document.querySelector('div [*|name=value]') did not throw
PASS: document.querySelectorAll('div [*|name=value]') did not throw
PASS: document.body.webkitMatchesSelector('div [*|name=value]') did not throw
PASS: document.querySelector('div [|name=value]') did not throw
PASS: document.querySelectorAll('div [|name=value]') did not throw
PASS: document.body.webkitMatchesSelector('div [|name=value]') did not throw
